Camera HAL: workaround for QCOM legacy camera parameters on KitKat
On Qualcomm legacy devices, the following camera parameters exist:
contrast-max, saturation-max, sharpness-max. However, newer camera
libraries name these parameters differently (max is the prefix,
not suffix).

This is patched in cameraservice, but in
frameworks/base/core/java/android/hardware/Camera.java,
the parameters are still hardcoded to max-contrast, max-saturation
and max-sharpness.

Workaround this by copying values in the camera HAL (or if
they don't exist, set to -1 so that the camera will still function).

Change-Id: I8e3e4e1b7a669e060e08018f02c9083dd3fffce8